fix: remove dependency of bounding box for visualizer
