+++ SplineMiner/Game/Debug/CartDebugManager.cs
@@ -3,6 +3,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using SplineMiner.Game.Cart;
 using SplineMiner.Game.Debug.Interfaces;
+using SplineMiner.Game.Debug.Visualizers;
 
 namespace SplineMiner.Game.Debug
 {
@@ -11,7 +12,8 @@ namespace SplineMiner.Game.Debug
     /// </summary>
     public class CartDebugManager
     {
-        private readonly List<ICartDebugVisualizer> _visualizers = [];
+        private readonly List<ICartDebugVisualizer> _cartVisualizers = [];
+        private readonly List<IDebugVisualizer> _boundingBoxVisualizers = [];
         private bool _isEnabled = true;
 
         /// <summary>
@@ -24,24 +26,45 @@ namespace SplineMiner.Game.Debug
         }
 
         /// <summary>
-        /// Adds a visualizer to the debug manager.
+        /// Adds a cart-specific visualizer to the debug manager.
         /// </summary>
         /// <param name="visualizer">The visualizer to add.</param>
-        public void AddVisualizer(ICartDebugVisualizer visualizer)
+        public void AddCartVisualizer(ICartDebugVisualizer visualizer)
         {
-            if (!_visualizers.Contains(visualizer))
+            if (!_cartVisualizers.Contains(visualizer))
             {
-                _visualizers.Add(visualizer);
+                _cartVisualizers.Add(visualizer);
             }
         }
 
         /// <summary>
-        /// Removes a visualizer from the debug manager.
+        /// Adds a bounding box visualizer to the debug manager.
+        /// </summary>
+        /// <param name="visualizer">The visualizer to add.</param>
+        public void AddBoundingBoxVisualizer(IDebugVisualizer visualizer)
+        {
+            if (!_boundingBoxVisualizers.Contains(visualizer))
+            {
+                _boundingBoxVisualizers.Add(visualizer);
+            }
+        }
+
+        /// <summary>
+        /// Removes a cart-specific visualizer from the debug manager.
         /// </summary>
         /// <param name="visualizer">The visualizer to remove.</param>
-        public void RemoveVisualizer(ICartDebugVisualizer visualizer)
+        public void RemoveCartVisualizer(ICartDebugVisualizer visualizer)
         {
-            _visualizers.Remove(visualizer);
+            _cartVisualizers.Remove(visualizer);
+        }
+
+        /// <summary>
+        /// Removes a bounding box visualizer from the debug manager.
+        /// </summary>
+        /// <param name="visualizer">The visualizer to remove.</param>
+        public void RemoveBoundingBoxVisualizer(IDebugVisualizer visualizer)
+        {
+            _boundingBoxVisualizers.Remove(visualizer);
         }
 
         /// <summary>
@@ -53,9 +76,16 @@ namespace SplineMiner.Game.Debug
         {
             if (!_isEnabled) return;
 
-            foreach (var visualizer in _visualizers)
+            // Draw cart-specific visualizations
+            foreach (var visualizer in _cartVisualizers)
             {
                 visualizer.Draw(spriteBatch, cart);
+            }
+
+            // Draw bounding box visualizations
+            foreach (var visualizer in _boundingBoxVisualizers)
+            {
+                visualizer.Draw(spriteBatch, cart.BoundingBox);
             }
         }
     }
